move dto & models in one folder
