+++ src/models/pokemon.dto.ts
