+++ src/models/species.dto.ts
